apps/analytics/pyproject.toml
[project]
name = "marketing-analytics"
version = "0.1.0"
description = "Measurement agent for marketing performance"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "pydantic-settings>=2.0",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "pandas>=2.0",
  "anthropic>=0.39",
  "aiosqlite>=0.20",
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "facebook-business>=20.0",
  "Pillow>=10.1.0",
  "httpx>=0.27",
  "ruff>=0.6",
  "black>=24.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
